+++ src/ceos/command/filter/columnfilter.rs
@@ -163,10 +163,10 @@ mod tests {
         4 keep me\n";
         let (sender, _) = std::sync::mpsc::channel();
         let mut buffer = Buffer::new_from_string(sender, content, 2);
-        assert_eq!(content.len(), buffer.len());
+        assert_eq!(content.len() as u64, buffer.len());
         assert_eq!(5, buffer.line_count());
         filter.execute(&mut buffer);
-        assert_eq!(content.len() - 8, buffer.len());
+        assert_eq!(content.len() as u64 - 8, buffer.len());
         Ok(())
     }
 }
