+++ src/ceos/command/filter/linedrop.rs
@@ -104,7 +104,7 @@ mod tests {
     fn test_filter_prefix() -> Result<(), ()> {
         let (sender, _) = std::sync::mpsc::channel();
         let mut buffer = Buffer::new_from_string(sender, CONTENT, 2);
-        assert_eq!(CONTENT.len(), buffer.len());
+        assert_eq!(CONTENT.len() as u64, buffer.len());
         assert_eq!(5, buffer.line_count());
         let filter = LineDrop::try_from("l ..2")?;
         filter.execute(&mut buffer);
@@ -119,7 +119,7 @@ mod tests {
     fn test_filter_range() -> Result<(), ()> {
         let (sender, _) = std::sync::mpsc::channel();
         let mut buffer = Buffer::new_from_string(sender, CONTENT, 2);
-        assert_eq!(CONTENT.len(), buffer.len());
+        assert_eq!(CONTENT.len() as u64, buffer.len());
         assert_eq!(5, buffer.line_count());
         let filter = LineDrop::try_from("l 3..")?;
         filter.execute(&mut buffer);
@@ -134,7 +134,7 @@ mod tests {
     fn test_filter_suffix() -> Result<(), ()> {
         let (sender, _) = std::sync::mpsc::channel();
         let mut buffer = Buffer::new_from_string(sender, CONTENT, 2);
-        assert_eq!(CONTENT.len(), buffer.len());
+        assert_eq!(CONTENT.len() as u64, buffer.len());
         assert_eq!(5, buffer.line_count());
         let filter = LineDrop::try_from("l 2..4")?;
         filter.execute(&mut buffer);
