+++ src/ceos/command/search.rs
@@ -167,7 +167,7 @@ mod tests {
         4 keep me\n";
         let (sender, _) = std::sync::mpsc::channel();
         let mut buffer = Buffer::new_from_string(sender, content, 2);
-        assert_eq!(content.len(), buffer.len());
+        assert_eq!(content.len() as u64, buffer.len());
         assert_eq!(4, buffer.line_count());
         filter.execute(&mut buffer);
         assert_eq!(2, buffer.line_count());
